+++ weather_forecast/src/forecast/mod.rs
@@ -1,6 +1,11 @@
+extern crate dotenv; // .env file (API_KEY)
 use chrono::prelude::*;
+use dotenv::dotenv;
 use reqwest::Client;
 use std::error::Error;
+use std::io;
+
+use crate::view::print_forecast;
 
 pub struct WeatherData {
     pub city: String,
@@ -28,20 +33,6 @@ pub fn get_cities() -> Vec<&'static str> {
         "Dinant",
         "Nivelles",
     ];
-}
-
-/**
- * Prints the forecast based on the given weather data.
- */
-pub fn print_forecast(forecast: &WeatherData) {
-    println!("Weather in {} :", forecast.city);
-    println!("\t- Time: {}", forecast.time);
-    println!("\t- Weather: {}", forecast.weather);
-    println!("\t- Description: {}", forecast.description);
-    println!("\t- Temperature: {:.2} °C", forecast.temperature); // 2 decimals
-    println!("\t- Humidity: {} %", forecast.humidity);
-    println!("\t- Wind speed: {:.2} m/s", forecast.wind);
-    println!();
 }
 
 /**
@@ -234,6 +225,134 @@ pub async fn get_weather_data_tomorrow(
         Err(err) => {
             println!("Error in the request: {:?}", err);
             return Err(err.into());
+        }
+    }
+}
+
+/**
+ * Reads the input from the user and loops until a valid city from the list is given.
+ */
+pub fn read_city_input(list_cities: Vec<&str>) -> String {
+    let mut input = String::new();
+    println!("\nSelect a city to display its weather forecast: ");
+
+    loop {
+        input.clear();
+        io::stdin().read_line(&mut input).unwrap();
+        println!("");
+        let selected_index: Result<usize, _> = input.trim().parse();
+
+        match selected_index {
+            Ok(index) => {
+                if index > 0 && index <= list_cities.len() {
+                    return (&list_cities[index - 1]).to_string();
+                } else {
+                    println!(
+                        "Please enter a valid number between 1 and {} : ",
+                        list_cities.len()
+                    );
+                }
+            }
+            Err(_) => {
+                println!(
+                    "Please enter a valid number between 1 and {} : ",
+                    list_cities.len()
+                );
+            }
+        }
+    }
+}
+
+/**
+ * Displays the weather forecast for all 10 Belgian cities.
+ */
+pub async fn task1(client: &Client, api_key: &str) {
+    let list_cities: Vec<&str> = get_cities();
+
+    for city in &list_cities {
+        let forecast_result = get_weather_data(&client, &api_key, city).await.unwrap();
+        print_forecast(&forecast_result);
+    }
+}
+
+/**
+ * Displays a list of city to choose from, to then display its weather forecast
+ */
+pub async fn task2(client: &Client, api_key: &str) {
+    // Displaying list of cities
+    println!("\n[List of cities]");
+    let list_cities: Vec<&str> = get_cities();
+
+    for (index, city) in list_cities.iter().enumerate() {
+        println!("  {} - {}", index + 1, city);
+    }
+
+    // Reading user input and retrieving city
+    let selected_city = read_city_input(list_cities);
+    println!("You selected: {}\n", selected_city);
+
+    // Displaying weather forecast
+    dotenv().ok();
+
+    let forecast_result = get_weather_data(&client, &api_key, &selected_city)
+        .await
+        .unwrap();
+
+    print_forecast(&forecast_result);
+}
+pub async fn task2_new_city(client: &Client, api_key: &str) {
+    let mut input = String::new();
+
+    loop {
+        println!("\nWrite down a city to display its weather forecast: ");
+        input.clear();
+        io::stdin().read_line(&mut input).unwrap();
+        println!("");
+        let selected_city = input.trim();
+        let forecast_result = get_weather_data_unknown(&client, &api_key, &selected_city).await;
+
+        match forecast_result {
+            Ok(forecast) => {
+                print_forecast(&forecast);
+                break;
+            }
+            Err(err) => {
+                println!("{}", err);
+                println!("The city doesn't exist !");
+            }
+        }
+    }
+}
+
+pub async fn task2_tomorrow(client: &Client, api_key: &str) {
+    let mut input = String::new();
+
+    loop {
+        println!("\nWrite down a city to display its weather forecast: ");
+        input.clear();
+        io::stdin().read_line(&mut input).unwrap();
+        println!("");
+        let selected_city = input.trim();
+        let forecast_result = get_weather_data_tomorrow(&client, &api_key, &selected_city).await;
+
+        match forecast_result {
+            Ok(forecast) => {
+                // Print each forecast
+                println!("[TODAY]");
+                print_forecast(&forecast[0]);
+
+                println!("[TOMORROW]");
+                print_forecast(&forecast[1]);
+
+                println!("[AFTER TOMORROW]");
+                print_forecast(&forecast[2]);
+
+                break;
+            }
+            Err(err) => {
+                println!("{}", err);
+                println!("The city doesn't exist !");
+            }
         }
     }
 }
